paxmon: loader refactoring, remove vehicle order code
