+++ modules/paxmon/include/motis/paxmon/capacity.h
@@ -17,7 +17,7 @@
 #include "motis/core/schedule/time.h"
 
 #include "motis/paxmon/capacity_data.h"
-#include "motis/paxmon/vehicle_order.h"
+#include "motis/paxmon/trip_formation.h"
 
 namespace motis::paxmon {
 
@@ -45,20 +45,17 @@ using trip_capacity_map_t = std::map<cap_trip_id, std::uint16_t>;
 using category_capacity_map_t = mcd::hash_map<mcd::string, std::uint16_t>;
 using vehicle_capacity_map_t =
     mcd::hash_map<std::uint64_t /* UIC number */, vehicle_capacity>;
-using trip_vehicle_map_t =
-    mcd::hash_map<boost::uuids::uuid, trip_vehicle_order>;
+using trip_formation_map_t = mcd::hash_map<boost::uuids::uuid, trip_formation>;
+using trip_uuid_map_t = mcd::hash_map<primary_trip_id, boost::uuids::uuid>;
 
 struct capacity_maps {
   trip_capacity_map_t trip_capacity_map_;
   category_capacity_map_t category_capacity_map_;
-  vehicle_capacity_map_t vehicle_capacity_map_;
-  trip_vehicle_map_t trip_vehicle_map_;
-};
 
-std::size_t load_capacities(schedule const& sched,
-                            std::string const& capacity_file,
-                            capacity_maps& caps,
-                            std::string const& match_log_file = "");
+  vehicle_capacity_map_t vehicle_capacity_map_;
+  trip_formation_map_t trip_formation_map_;
+  trip_uuid_map_t trip_uuid_map_;
+};
 
 std::pair<std::uint16_t, capacity_source> get_capacity(
     schedule const& sched, light_connection const& lc,
