+++ modules/paxmon/src/loader/motis_journeys/journey_access.cc
@@ -1,4 +1,4 @@
-#include "motis/paxmon/loader/journeys/journey_access.h"
+#include "motis/paxmon/loader/motis_journeys/journey_access.h"
 
 #include "utl/enumerate.h"
 #include "utl/verify.h"
